Fixes a compile warning.
git-svn-id: http://svn.mythtv.org/svn/trunk@7089 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/guidegrid.cpp
@@ -722,9 +722,9 @@ void GuideGrid::fillProgramRowInfos(unsigned int row)
     }
 
     int chanNum = row + m_currentStartChannel;
-    if (chanNum >= (int)m_channelInfos.size())
-        chanNum -= (int)m_channelInfos.size();
-    if (chanNum >= m_channelInfos.size())
+    if (chanNum >= (int) m_channelInfos.size())
+        chanNum -= (int) m_channelInfos.size();
+    if (chanNum >= (int) m_channelInfos.size())
         return;
 
     if (chanNum < 0)
